Factor out source listing tests (#625)
